feat: 여행 일정 수정 api 추가

+++ src/main/kotlin/com/brtrip/path/domain/PathPlace.kt
@@ -20,5 +20,5 @@ class PathPlace(
     var place: Place,
 
     @Column(name = "sequence", nullable = false)
-    var sequence: Long
+    var sequence: Int
 ): BaseEntity()
